Calculate last test in a file differently
Recent changes in bats-core, in order to implement setup and
teardown facilities, have changed how internal test names work.

https://github.com/bats-core/bats-core/issues/39 (and others)

Because if this we cannot use those test names to detect
the last test in a file anymore.

So we need to rely exclusively on BATS_TEST_NUMBER, that
continues working ok.

That's all we do here and it should make the tests to be back to pass.