Add renamed classes following MDL-75716
In MDL-75716, unit test class names were updated to match file names. As
a result the parent class for a couple of tests was changed and must be
added to the list of known parent classes.

The old location is not removed as they may still be legitimately used
in older versions of Moodle.