Move from progressIssue to transitionIssue
The progressIssue action was deprecated in Jira CLI 6. And
replaced by better transitionIssue.

Also, move away from --custom , deprecated in Jira 10 and
instead use --field, that supports both standard and custom
fields (and can be used multiple times, one for every field.

Note that, because of https://jira.atlassian.com/browse/JRASERVER-61539
is still impossible to use the correct updateIssue action, because it's
not allowed to update fields not available in the current screen.

So we need to continue using those "CI Global Self-Transition"
ghost transitions (from -> to the same status), using a hidden
screen having ALL the fields, via the transitionIssue action
in order to get those fields updated. Sigh.