Provide compatibility with Node 18 and a few more little things
We need to support Node 18 from the CI tools (see MDL-75012). So
this commit provides that, basically.

The main difference is that the `npm bin` command is not available
anymore, and that's what we used both for:

1) Search if a given binary has been installed (say, grunt-cli...)
2) Launch a number of binaries (grunt, eslint, gherkin-lint....)

So here we are geting rid of all those uses, changing them by:

1) `npm list` to verify if a package is already installed.
2) `npx` to launch the binaries.

Also, we have removed some very-old conditional blocks in the code, that
were providing compatibility with very, very old Moodle branches, before
we used grunt and friends.

Finally, we are now installing stylelint-checkstyle-formatter if
it's detected that core has not installed it. After all, it's a CI
dependency, not a core one, so maybe the more correct thing is to
delegate its installation to the CI / testing tools.

And, basically, that is. I've tested it with both Node 16 (current
master) and Node 18 (MDL-75012) and seems to be working ok.

Going to run some tests now...