Ensure the uniqueness of paths being preserved
Until now we were preserving all the files in any directory like:

- */.git/*
- */work/*
- */node_modules/*

And it has worked perfectly (locally, CIs, Travis...) because none
of them used any of ".git", "work" or "node_modules" directories
within the workspace where the utility is being executed.

But, with GHA, all the workflows are executed under the directory:

/home/runner/work/repo...

And that was leading to ALL files being preserved, because all them
match the parent "work" directory. So now we are more accurate
about which directories we want to preserve, only:

- workspace/.git/*
- workspace/work/*
- workspace/node_modules/*

Avoiding any casual match with parent directories.