Ensure that short hashes are, at very least 12 chars long
Till now, given the moodle.git repo size, short hashes had
a length of 10 chars, but we are arriving to the point they
are jumping onto the 11 chars wagon (some repos already show
them with the new length).

As far as various tests are dependent of the short hashes, we
are going to use a fixed --abbrev=12 (minimum to show) where
needed. That way results will be consistent and not depending
of git heuristics to calculate the size. This should keep covered
some good years.