New criterion to process (ToBiC) those issues waiting for CLR
The  is exactly the same than the Waiting for Integration (WFI) one,
but applied to the Waiting for CLR status (WFCLR).

Note that, right now, it won't effectively, run many jobs, because
right now all issues first land to WFI before being manually moved
to WFCLR. And that usually happens hours later, so the existing WFI
job already has processed them.

But in the future, when we start automating the IR/CLR triaging, there
will be more and more issues in WFCLR status than now.

This corresponds to point #3 of the Associated Tasks in:

https://tracker.moodle.org/browse/MDLSITE-6610