Add the @javascript or ~@javascript tag based on the job
Normally the ci jobs handle this automatically and, when
no tags are specified, depending on he browser (goutte
for non-js and others for js) both the @javascript and
~@javascript tags are added.

But when an explicit tag is specified, it takes precedence
and the ci jobs don't add/remove anything.

Hence, we have to, explicitly add them here when a tag
has been specified.