Don't run the grunt check if there aren't css/js changes
By far, the grunt check is one of the slowest ones, accounting
a very high % of the total time. While this will improve in
the future (for example when bumping to new nodejs 18 stuff
we detected 10x times improvements building javascript), still
this little hack  is good, causing the job to finish
really quicker when there isn't anything to be checked by grunt.