Ensure that we always generate the grunt and shifter sections
No matter that now we aren't running the grunt check always, we
want the grunt section (and the shifter one) to always exist
in the final report.

Also, amend the gherkin-lint tests because now they don't run
grunt (as expected), so the 3 errors are only reported by
gherkin, not by grunt.

Finally, apply for a pre-existing (since inception) bug that was
causing test not to merge (in integration.git) against the
desired base. That may lead to confusion because moodle.git merges
were being done correctly (with the commit specified in the tests
being respected as base for the merge).