Make all the "holding" actions to apply both for IR & CLR
Now, all the actions related to holding issues:
- After freeze, any new feature or improvement.
- Last week, any issue type.
- While on-sync, any new feature and improvement.

Are managed both for issues awaiting for IR and awaiting
for CLR, so the same outcome (hold issues) happen
exactly the same for both.

Note that we'll have to modify the manual release process slightly
to remember to unhold issues from both queues, that will be
documented @ https://moodledev.io/general/development/process/release
once this is merged and working.