Fix a recently introduced / modified bogus test
One of the tests was comparing dynamically with 402_STABLE
branch and that broke today (October 9th 2023) with a new
release happening in that branch.

So we are making the test to be consistently performing
always the same fixed upgrade, from v4.2.2 to v4.3.0 (patched).

That way future upgrades won't affect it, as far as the 2 versions
used have been already released and tagged and won't change ever.