New job to be able to keep two branches on sync
This job just gets two branches from a given remote (default
to origin) and ensures that the target branch is kept 100%
on sync with the source one.

It detects if commits have been sent to the target branch
by mistake and also, is able to detect rewrites and process
them properly in the target branch.

Covered with tests.