Move from participants to JQL Tricks commentedByUser()
The original query was working wrongly sometimes, because of
Jira re-indexing problems. And that was leading to some ToBiC
jobs to be relaunched, when they should not.

We have confirmed that, after a re-index, the original query
is back to work, but that's not nice-enough outcome.

So we have moved to the already installed JQL Tricks query
that seems to be immune to the indexing problems and properly
keeps out to all the issues where ToBiC already has participated
(by adding a comment).

Note that it has been checked that the JQL Tricks extension is
also available for Jira Cloud (with a change in the syntax), but
will work there, so it's ok if we add this new use here.

Ref: https://www.j-tricks.com/jql-tricks-cloud-migration.html