Validate release date on continuous queues manager
Before performing the job tasks, we need to make sure that the release
date is properly configured. This can be simply done by making sure
that the current date is not well past the end of the on-sync date,
which is calculated by adding 4 weeks to the configured release date.
(Normally, on-sync ends 2 weeks after the release, but it's set to 4
weeks in this patch just in case we have an unusually longer on-sync
period.)

This commit also moves the parameter validation part to its own
function, so it can be tested on its own.

Additional validation to ensure that the `lastweekdate` is not set
after the `releasedate` has also been added.