Fix edge case when diff changes were exactly "++ " or "-- "
In those cases, sometimes, that was leading to the lines
to be recognised as a new file incorrectly, causing various
troubles later in the process.

Now we look for the two lines in a row, and only when they
have something looking like a path:

- /dev/null
- /path/to/something

Covered with test containing the edge cases.