Set the correct phpcs runtime options for plugins
Until now we were completely excluding some sniffs to
avoid them to be run.

While that's correct, better if we use the config
options created for that.

Note that, because we have to pass empty ('') strings
we need to handle all the arguments as an array,
so shell doesn't do strange things and the binary
being called do receive everything (empty strings included)
as proper arguments.

Trying to pass the information as a string doesn't work ok
when handling empty strings.