Replace many assert_files_same() by simpler assert_file_matches()
We have left a couple of tests in the remote_branch_checker tests
still using the whole-file xml fixture, so we can verify that the
file structure (smurf file) is kept stable. All the rest of tests
have been moved to the regext based fixtures.

With that, changes within the remote_branch_checker (CiBoT) should
be easier to maintain only affecting to a couple of files, not to
all them.