For 4.5dev and up, let's detect any change to upgrade files
This includes both:
 - upgrade.txt files, that shouldn't be modified ever.
 - UPGRADING.md files, which are 100% managed by mdlrelease.

Note that, if some day we change something in the tool generating the
UPGRADING.md files and we apply it as part of an issue, the check will
report it as an error, but sure that we can consider that a false
positive and proceed to integrate it.

In the other side, 99% of editions won't be legit, hence real positives.