Better detect checked stuff + force push to prececker repo
To better understand what has been already checked now we
are setting, as part of the branch names, both:
- The base commit in the upstream branch where the patch has
  been merged.
- The patch HEAD commit that is being merged.

That way, in the future, we can detect better which combinations
have been already checked.

Also, for now, let's --force the push to the prechecker repo, so
if something is checked twice (same base+patch commits), it's just
overwritten. Some day we may change this behaviour.