Change to yyy/mm/dd format everywhere
Previous format used (MMM) was locale-dependent, with some
hosts returning "Jun" and others returning "June". Maybe it
is a Java VM thing or a OS thing, but it was happening, leading
to problems when comparing dates (as strings).

So we are moving to locale-independent format, and also using
year/month/day because that's the format we need in queries
and makes everything easier to read.