Accept any _testcase as valid
Historically we had a blur between `_test` and `_testcase`, but in
recent versions we have largely solved this. We now expect:
- every test class name to be suffixed with `_test`
- every test class to be final
- every test class to extend a testcase with class name suffixed `_testcase`
- every testcase class to be abstract

Given these parameters, which have been in place for some time and are
checked with phpcs, we can move away from a hardcoded list of accepted
testcases, and move towards a regular expression based on the name.

There are still some uses of the legacy names, but we can reduce them
and support any _testcase class.