Ensure that must-fixes and mdlqa issues in CLR are not held
Adding conditions for the CLR part (filter=23329) of the A3b script so
that mdlqa and must-fixes don't get held, especially during the last
week before the release.

The conditions are not applied to the ones in IR (filter=14000) because
these are already being properly handled in A1 and A2 functions.